Add types and checked casts, update codebase

+++ include/shclog/pause.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
+#include "shclog/types.hpp"
 #include <emmintrin.h>
 
 namespace shclog::pause {
-template <int N> inline void mm_pause() {
-    for (int i = 0; i < N; ++i)
+template <usize N> inline void mm_pause() {
+    for (usize i = 0; i < N; ++i)
         _mm_pause();
 }
 } // namespace shclog::pause
